Don't use switch-case statements for resource IDs
We plan to migrate users to have IDs in the R class
not be constants, as it allows us to make some large
build performance improvements.

However this is a source-level breaking change
as the IDs in the R class can no longer be used
as the cases in a switch statement.

This change updates the data binding generated code
to use a chained if instead.

Test: DataBindingTest
Bug: 126376860
Change-Id: I43db1adf0f804c1a297a2aa1b787027bd6c52f81