Detect duplicate ids in layouts
Data Binding was only verifying view ids between multi-config layouts because
lint already warns for duplicate ids in views. On the other hand, not detecting
this error during layout processing causes compilation to continue and fail with
an unrelated error later on.
This CL fixes that issue and detects the problem during base layout generation.

I've also changed base layout generator to not to eat errors. I'll make the
necessary changes on the gradle side to report them with the right encoding.

Bug: 119645762
Test: SimpleCompilationTest
Change-Id: Ib0939e9c160ada5337fb6703205d3140f446bede