Don't compute relative paths for out-of-project resources.
To ensure cacheability of the Android Gradle plugin's MergeResources
task, in a previous CL, we used relative paths for resource files,
assuming that they were located within the project directory.

This won't make sense and won't work for cases when a resource file is
located outside of the project directory, e.g. if the user adds a custom
resource source set (bug 128579779).

In that case, we will need to keep using the absolute path of the
resource file, even though it breaks cacheability.

Bug: 128579779
Test: Updated RelativizableFileTest
Change-Id: I976871d339bb68ee60253c803ad97770fae05d9e