Minimize impact of exception-handling code for missing views
The existing generated code already de-duplicated the exception message strings. This change de-duplicates the bytecodes required to create the message and throw the exception. It also has a pleasant side-effect of changing the bytecode to remove all jumps in the common case (when required views are all present).

For each view, the validation code changes from

    if (name == null) {
      throw new NullPointerException("Missing required view with ID: ".concat("name"));
    }

to

    if (name == null) {
      missingId = "name";
      break missingId;
    }

The whole "body" of view lookup is now surrounded with a labeled block with a local and single exception handling code.

    String missingId;
    missingId: {
      // View lookups here...
      return new MainBinding(...);
    }
    throw new NullPointerException("Missing required view with ID: ".concat(missingId));

This results in very lean bytecode.

    [0001dc] MainBinding.bind:(Landroid/view/View;)LMainBinding;
    0000: sget v0, Lcom/example/R$id;.name:I
    0002: invoke-virtual {v2, v0}, Landroid/view/View;.findViewById:(I)Landroid/view/View;
    0005: move-result-object v0
    0006: check-cast v0, Landroid/widget/TextView;
    0008: if-eqz v0, 001d

    000a: sget v0, Lcom/example/R$id;.email:I
    000c: invoke-virtual {v2, v0}, Landroid/view/View;.findViewById:(I)Landroid/view/View;
    000f: move-result-object v2
    0010: check-cast v2, Landroid/widget/TextView;
    0012: if-eqz v2, 001a

    0014: new-instance v2, Lcom/example/MainBinding;
    0016: invoke-direct {v2}, Lcom/example/MainBinding;.<init>:()V
    0019: return-object v2

    001a: const-string v2, "email"
    001c: goto 001f

    001d: const-string v2, "name"

    001f: new-instance v0, Ljava/lang/NullPointerException;
    0021: const-string v1, "Missing required view with ID: "
    0023: invoke-virtual {v1, v2}, Ljava/lang/String;.concat:(Ljava/lang/String;)Ljava/lang/String;
    0026: move-result-object v2
    0027: invoke-direct {v0, v2}, Ljava/lang/NullPointerException;.<init>:(Ljava/lang/String;)V
    002a: throw v0

As you can see the bytecode executes linearlly without branching in the case where no views are null from 0000 to 0019. Upon finding an unexpectedly null view, a jump will occur to initialize the name in v2 which then jumps to 001f where the consolidated exception-handling code is.

This has a net savings of approximately 20 bytes per view. There's still a fixed 24 bytes for the consolidated exception handling and 6 bytes per view to load the ID name and jump to exception handling.

Bug: 128314294
Test: manual
Change-Id: I5ec1042b80d3e18b0b12beca1eaab1e8a9babc92