Re-introduce ViewBinding constructor w/ DataBindingComponent
This CL fixes a bug where libraries compiled with AGP 3.3 and before cannot be loaded
in AGP 3.4 because the constructer has been removed. The only fix is to re-compile
the library with a newer version of AGP which is not always feasible.

This CL fixes the problem by re-adding the old constructor back.

Bug: 122936785
Test: MultiModuleTestApp#OldVersionDependencyTest
Change-Id: Ie5aed4ca7e8df2225c186baf04ff956996507888