Third attempt: Fix thread safety issue in LibTypes class.
The LibTypes class has a call to Jetifier's
ConfigParser.loadDefaultConfig(), which then calls
Class.getResourceAsStream(...), and that method is not thread safe.

The use of Class.getResourceAsStream(...) is not thread safe when the
class is loaded by a URLClassLoader, as there is a bug in the
implementation of URLClassLoader.getResourceAsStream(...):
https://bugs.openjdk.java.net/browse/JDK-8205976.

To fix that, we now use Class.getResource(...).openStream() instead,
which should be equivalent to Class.getResourceAsStream() but doesn't
have the thread safety issue above.

The fix should be in the Jetifier code, but we want to release this fix
early for developers to try out, so we implement this fix in data
binding first. We will remove this fix in data binding later once we
update data binding to use a newer Jetifier version that has the fix.

Bug: 137929327
Fixes: 120277395
Test: Existing tests + Manually verified on a repro project
Change-Id: I54fe25df87bdf0f8a9c1199df403b60b98301871