Fix generic discovery when xml does not have it
This CL fixes a bug where if there is a class declareted as Foo<T>
but the XML declarates it as Foo (without type args), we would
fail to resolve any type inside it.
e.g.
class Foo<T> {
    List<String> getList()
}

would resolve to getList returning just a List.

This CL fixes it by separating how we turn a class into code based
on context (whether as a variable declaration or regular code, e.g.
to access static fields)

It is not super clean but there is a lot of complicated scenerios
so I tried to fix it w/o breaking anything else or changing too
much logic.

Also added infra to compiler tests to be able to create tests that
run in the processing environment to make future debugging easier.
We should ultimately move all JavaModel tests into that and get
rid of JavaModel implementation (if not used by studio).

Bug: 139738910
Test: AnnotationModelTest, GenericInterfaceTest, GenericAdapterTest
Change-Id: I2ea4ad74596fd6145f8dd4665dd550fd6e36f0d6