Allow setters/adapters in <include tags for binding
This CL fixes a missing feature/bug where if an <include tag resolves to a Binding
class, it would not support any binding adapters but worked perfectly if included
layout is not a binding layout.

This CL moves it to use a common architecture where if a binding target is a view
binding, we'll try to unroll it to its getRoot method if a setter cannot be found
in the binding class.

Bug: 133390436
Test: IncludeTagTest.java
Change-Id: I26e5b458303a185f1f3fa42c3e4fd0790d601dd7