Ensure all strings are capitalized w/ US locale
This CL fixes a bug in data binding where in some places,
capitalization did not specify a locale, which created
problems on windows machines that are set to Turkish since
turkish i is still an i when capiltalized (not an I).

Surprisingly, it works fine on linux machines while fails
to compile on windows machines but i was able to create a
test case which broke data bindng when two way binding is used.

This CL is a bit risky if some turkish user depended on this
behavior on a linux machine and accessed those fields in
their code but i think it is better to fix this way since that
is what kotlin does as well.
https://github.com/JetBrains/kotlin/blob/64e85e8a0c7b0c0392296bcd8b3dd9335bbf1ef4/core/descriptors/src/org/jetbrains/kotlin/name/NameUtils.kt

Bug: 141925966
Bug: 118943890
Test: NonEnglishLocaleTest in compilation tests

Change-Id: If879c52639530431a0710537136fc934f421b489