Remove explicit use of strings for view error handling
Instead, re-use the ID value to look up the name from the resource table only in the event of failure. This removes string references to view names when obfuscating. (It's worth noting that this does not actually prevent leaking the name entirely since they're still present in the resources.arsc table which needs separately obfuscated).

This results in a trivial increase in Dalvik bytecode for binder classes with 1 or 2 bindings as a result of the getResources() and getResourceName() method calls. For classes with 3 or more bindings, however, the code is smaller due to the lack of loading the string constants and additional jump instructions.

Before:

    public static MergeBinding bind(@NonNull View rootView) {
      String missingId;
      missingId: {
        View one = rootView.findViewById(R.id.one);
        if (one == null) {
          missingId = "one";
          break missingId;
        }

        View two = rootView.findViewById(R.id.two);
        if (two == null) {
          missingId = "two";
          break missingId;
        }

        return new MergeBinding(rootView, one, two);
      }
      throw new NullPointerException("Missing required view with ID: ".concat(missingId));
    }

After:

    public static MergeBinding bind(@NonNull View rootView) {
      int id;
      missingId: {
        id = R.id.one;
        View one = rootView.findViewById(id);
        if (one == null) {
          break missingId;
        }

        id = R.id.two;
        View two = rootView.findViewById(id);
        if (two == null) {
          break missingId;
        }

        return new MergeBinding(rootView, one, two);
      }
      String missingId = rootView.getResources().getResourceName(id);
      throw new NullPointerException("Missing required view with ID: ".concat(missingId));
    }

Bug: 143160165
Test: ../base/bazel/bazel test //tools/base/build-system/integration-test/databinding:tests
Test: ../base/bazel/bazel test //tools/data-binding/...
Change-Id: I25524065b06dca540a74f9f5a1297ae60a63cf0e