Handle cases where upperbound of a TypeVariable might return null
This CL adds a safety to how we handle upper bounds.
There is a very interesting bug where if someone bumps the jetbrains
annotations to v15 from 13 (which is depended upon by kotlin), the
compiler weirdly returns null for the upper bound of an annotated type
variable.
Unfortunately, there is no test as I couldn't figure out how to reproduce
this besides that interesting instance. I've looked into the jars for v15
and v13 and nullable class is there both.

I've also added more logging to such failures where we receive an exception
that we don't expect. This will help with errors in the future where we'll
at least know which layout caused the error hence we can diagnose it better.

Sample error output when invoked from IDE:
https://paste.googleplex.com/6688355217571840
Sample error output when invoked from command line:
https://paste.googleplex.com/5133349121163264

FYI this is still an exception not a deferred thing as it is really unexpected
hence no reason to defer execution and lose the cause of the exception.

Bug: 144300600
Test: unfortunately manual but I've verified the repro in the bug is fixed or it
also prints the layout name w/ the stack trace if the exception happens (before
the fix)

Change-Id: Ie42f24703f4dca134e7848a3eac9b7b7c45f55eb