Debug data binding compilationTests hangs
Data binding compilation tests launch Gradle builds in a separate
process and sometimes hang when waiting for the process to complete.
This happens on build server, but I've also seen it locally.

To debug this, this CL will set a timeout for the wait, and if the
timeout happens, print out the output and error logs from the process
so far.

Also make sure we kill the unfinished processs.

Bug: 148496072
Bug: 148598093 (maybe also 146217083)
Test: Existing tests
Change-Id: I859361bbd55f32ff49d8235c015511e763a90f53