Clear static state in Scope when reporting errors
As soon as assertNoError() is invoked, deffered
exceptions can be removed. This makes sure the next
time data binding runs it will not report the same
issues.

Previous mechanism relied on AGP to clear the exceptions
but that does not seems necessary as the exceptions
become obsolete once reported and LoggedErrorException
is thrown.

Bug: 154742527, 151860061
Test: existing
Change-Id: Ic70805e390f5034d7c61659ead7002f36b185669