Set max_workers to 1 in data-binding targets.
Per b/155772061, running gradle tasks with the default number of workers on Mac is possiblity deadlocking gradle and hang the build. So we try to set the max_workers to 1 and hopefully solve this problem. Build performance impact should be minimal (a local run of the targets did not show any slow down).

Test: presubmit tests
Bug: 155772061
Change-Id: Ibee0b46c6b08a769c5d6ffbfa096b6e8ef90e7bf