Fix JavaClassTest checks to be more robust against future JDK versions
Previously, we were verifying parsing by checking for specifically named
fields and functions, but some of those fields / methods have been
removed starting in JDK11.

Since we don't care about exact fields / methods, but just more that the
parsing code works, we instead check the properties of the fields /
methods for verification, which should be a lot more future proof.

Fixes: 161611073
Test: Only test code updated
Change-Id: I0d330e0247edd051206ad186afeaa929c7d3496e