Fix undefined generic assignability
This CL fixes a bug in data binding, or rather makes it more compatible
with java programming language.

We broke incomplete generic assignability in 3.5 timeframe where
we stopped supported assigning List<Foo> to List to fix some other
generated code. (b/139738910, I2ea4ad74596fd6145f8dd4665dd550fd6e36f0d6).

Unfortunately, it is a common patter to define such layouts where the
layout is not necessarily interested in the type variable.
Also, technically, java is completely fine with assigning List to List<String>
or List<String> to List.

To accomodate for that, this CL changes the isAssignable check in ModelClass
to also accept assignability if one of the types is a TypeVar (incomplete).

Bug: 152153710
Test: AnnotationModelAssignabilityTest.kt
Change-Id: Ic1766fbddb5bb3c1508c83c1fd28aa236645fa5f