Puts local antlr before platform antlr
The databinding module needs antlr to run so it includes it
as a module library. This makes anltr be bundled in the plugin
lib directory. However the platform offers another antlr library
that is bundled with studio.

This change makes the development/test classpath closer to production,
and now the tests can pass, but this could potentially be an issue
down the road having two libraries with different versions loaded
in the plugin. If another module uses the platform antlr, at runtime
is going to receive a different version.

Test: this is now running in studio_linux
Bug: 156497102
Change-Id: I0e670e1c5c1bf912b6ae606afbcc7e0ca46199f4