Handle type suffixes in the parser
This CL fixes a bug in data binding where if someone uses long type suffix,
we would still consider it as an integer.

I had to do it manually instead of from the grammar as it is declared as a
fragment, which makes it a non-token. I didn't want to change the grammar
to reduce the risk of impact for this change.

I've also added tests for hex, binary and float even though they are not
affected by this bug.

Also, unrelated to this CL but i needed to update mockito in test app
to make some mocking tests pass.

Bug: 167976837
Test: TypeSuffixTest
Change-Id: I65c6ebf4a013219a0ac12c0de672cc712ecea494