Keep a weak reference to LifecycleOwner in observers
WeakListeners in data binding are designed to hold on a strong
reference to the observed object but weak reference to the binding
it self so that we don't leak the binding.

Unfortunately, live data observers were keeping a strong reference
to the lifecycle owner, which meant that if binding is GC'ed and
we don't get a chance to process weak references, we might leak the
lifecycleowner.

Ideally, we wouldn't want to use this ReferenceQueue that cleans up
weak listeners but aside from that queue, we don't have a good
notification (on the main thread) that the binding can never be used
anymore so I decided to keep it for now.

Bug: 176886060
Test: existing tests, also adding tests that clear lifecycle
Change-Id: Ia72639571cf6aef2769de20383ee400ea20f67ba