Skip root view while finding child views
When a layout binding is included inside another layout, the include tag
might be setting a different id. If that id conflicts with one of the
children in the included binding, it breaks the generated Binding class
as it finds the root view instead of finding a child when calling
findVewById.

This CL adds a helper class to run find view by id only on children. We
don't need it for the root view as we already use the argument that was
passed in.

Bug: 177013456
Test: IncludeIdCollusionTest
Change-Id: I9e0739936d99f96d0396f2ea34b3df91fa609752